requirements.txt
asgiref==3.3.4
Django==3.2.3
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
isort==5.8.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
reportlab==3.5.67
sorl-thumbnail==12.7.0
sqlparse==0.4.1
